# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Xoycoin Exchange",
    "description": "Métodos de la API de Xoycoin Exchange",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api-xoycoin.xellcoders.com",
      "description": "Xoycoin Exchange API"
    }
  ],
...
  "paths": {
...
    "/deposits_users/validate/xoy": {
      "post": {
        "tags": [
          "deposits_users"
        ],
        "summary": "Validate XOY user deposit",
        "description": "Valida el depósito (XOY) de un usuario.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/validate_xoy_body_1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transaction",
            "content": {
              "application/json": {
                "example": {
                  "hash": "0x3c5cc2f09da36257bb3eb37a193d71f7759d5eba114c22190c0acb2edd94740b",
                  "fee": 0.001
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/deposits_users/wallet/{walletId}": {
      "get": {
        "tags": [
          "deposits_users"
        ],
        "summary": "Get user deposits by wallet id",
        "description": "Obtiene los depósitos de un usuario mediante el id de la wallet proporcionado.",
        "parameters": [
          {
            "name": "walletId",
            "in": "path",
            "description": "Id of the wallet to return deposits",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array deposits",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepositUser"
                },
                "example": [
                  {
                    "id": "b3de0177-020a-11ec-8efe-0242ac110003",
                    "walletId": "2c7d0d20-1cb5-4b21-b7e6-161480293c53",
                    "hash": "0x7ce560fa06ca34f0edf0b8922ee480a92cd62c46f85b9d1ca6743975624556df",
                    "sender": "mrd2zdpLJrdT6uVJKK6qkFz9FcPorbqbkc",
                    "amount": "1.0",
                    "fee": "0.001",
                    "confirmations": 1,
                    "status": "PENDING"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: User not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/levels": {
      "put": {
        "tags": [
          "levels"
        ],
        "summary": "Update level",
        "description": "Actualiza el nivel de una empresa.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/levels_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Level object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level"
                },
                "example": {
                  "id": 1,
                  "companyId": "61c0ccd7-63b6-4071-af7d-b392e0dadb67",
                  "order": 1,
                  "description": "Level name"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "levels"
        ],
        "summary": "Create level",
        "description": "Crea el nivel de una empresa.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/levels_body_1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Level object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level"
                },
                "example": {
                  "id": 1,
                  "companyId": "61c0ccd7-63b6-4071-af7d-b392e0dadb67",
                  "order": 1,
                  "description": "Level name"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/levels/company/{companyId}": {
      "get": {
        "tags": [
          "levels"
        ],
        "summary": "Get levels by company id",
        "description": "Obtiene los niveles de una empresa mediante el id proporcionado.",
        "parameters": [
          {
            "name": "companyId",
            "in": "path",
            "description": "Id of the company to return levels",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Levels array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level"
                },
                "example": [
                  {
                    "id": 1,
                    "companyId": "61c0ccd7-63b6-4071-af7d-b392e0dadb67",
                    "order": 1,
                    "description": "Level description"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Company not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/limits": {
      "put": {
        "tags": [
          "limits"
        ],
        "summary": "Update limit",
        "description": "Actualiza un límite.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/limits_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Level object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level"
                },
                "example": {
                  "id\"": 1,
                  "levelId": 1,
                  "type": "MXN",
                  "deposit": "100.0",
                  "withdrawal": "100.0"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "limits"
        ],
        "summary": "Create limit",
        "description": "Crea un límite.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/limits_body_1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Level object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Level"
                },
                "example": {
                  "id": 1,
                  "levelId": 1,
                  "type": "MXN",
                  "deposit": "100.0",
                  "withdrawal": "100.0"
                }
              }
            }
          },
          "401": {
            "description": "No token provided"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/limits/level/{levelId}": {
      "get": {
        "tags": [
          "limits"
        ],
        "summary": "Get limits by level id",
        "description": "Obtiene los límites de un nivel mediante el id proporcionado.",
        "parameters": [
          {
            "name": "levelId",
            "in": "path",
            "description": "Id of the level to return limits",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Limits array",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Limit"
                },
                "example": [
                  {
                    "id": 1,
                    "levelId": 1,
                    "type": "FIAT",
                    "deposit": "100.0",
                    "withdrawal": "100.0"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "No token provided"
          },
          "500": {
            "description": "ERROR: Level not found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Level": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "companyId": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Limit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "levelId": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "deposit": {
            "type": "string"
          },
          "withdrawal": {
            "type": "string"
          }
        }
      },
...
      "DepositUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "walletId": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "sender": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "fee": {
            "type": "string"
          },
          "confirmations": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "validate_xoy_body_1": {
        "required": [
          "amount",
          "hash",
          "signature",
          "userId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          }
        }
      },
      "levels_body": {
        "required": [
          "description",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "levels_body_1": {
        "required": [
          "companyId",
          "description",
          "order"
        ],
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string"
          },
          "order": {
            "type": "integer"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "limits_body": {
        "required": [
          "deposit",
          "id",
          "type",
          "withdrawal"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "deposit": {
            "type": "string"
          },
          "withdrawal": {
            "type": "string"
          }
        }
      },
      "limits_body_1": {
        "required": [
          "deposit",
          "levelId",
          "type",
          "withdrawal"
        ],
        "type": "object",
        "properties": {
          "levelId": {
            "type": "integer"
          },
          "type": {
            "type": "string"
          },
          "deposit": {
            "type": "string"
          },
          "withdrawal": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}